+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 
 - `BitmovinSdkAdapter` didn't transition to the `pause` state on a pause event
+- `droppedFrames` was not collected
 
 ## v1.6.3