+++ CHANGELOG.md
@@ -7,10 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Development
 
-## Fixed
+### Fixed
 
+- `BitmovinSdkAdapter` didn't transition to the `pause` state on a pause event
 - `droppedFrames` was not collected
-
 
 ## v1.6.3