+++ CHANGELOG.md
@@ -7,9 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Development
 
+### Internal
+- `BitmovinAnalyticsConfig` is serializable now. `Context` is now passed seperately where needed. For compat the old constructors are still there.
+
 ### Fixed
 
-- `BitmovinAnalyticsConfig` is serializable now. `Context` is now passed seperately where needed.
 - `BitmovinSdkAdapter` didn't transition to the `pause` state on a pause event
 - `droppedFrames` was not collected